requirements.txt
django-extensions==3.1.3
djangorestframework==3.12.4
django-cors-headers==3.7.0